Don't use dynamic page size
This mostly reverts 08050dd5e3701b29e5628e1ec7d37c1cd6529c57 (#54)
because the page size of the host is not necessarily the same as the
page size of the binary. For a proper fix, we'll need some way to
determine the page size of the binary, but ELF doesn't readily provide
this information.

Fixes #69.