Add travis configuration to run CI on github
This travis CI configuration will automatically build and run the test
suite of patchelf whenever a commit or pull request is made on Github.
It uses the `trusty` container-based configuration to deploy patchelf and
build it on an Ubuntu 14.04 container, since the version of autoconf
required by patchelf is newer than that provided by the `precise` containers.