Fix endian issue when creating sectionsByOldIndex
sectionsByOldIndex was resized to hdr->e_shnum instead of
rdi(hdr->e_shnum). This omitted the endianness conversion and probably
only worked by accident, because the 16-bit LE->BE conversion results
in integers that are larger as long as there no more than 255 section
headers.

This would be a good usecase for std::transform, but I'm unsure
whether we want to bump requirements to build patchelf to C++ 2017.