Introduce Finally to help track section rewrites
Similarly to a concept used in NixOS/nix codebase, introduce the Finally
class that can be given a lambda to execute on destructor.

The desired effect is to execute some code on a scope leave, in this
case the method return.

This is used to make sure we only apply sectionRewrite once per
modification and can now support early returns in the code paths.