patchelf.cc: add comment explaining calculation for DT_MIPS_RLD_MAP_REL
This commit also includes a link to the best source I could find for
what DT_MIPS_RLD_MAP_REL is supposed to be, an exact copy-paste of the
text found there, and step-by-step algebra for deriving the formula
used by the code.  Hopefully this means that this is the end of the
DT_MIPS_RLD_MAP_REL headaches.